Android 7.0.0 release 1

Conflicts:
	cmds/pm/src/com/android/commands/pm/Pm.java
	core/java/android/app/WallpaperManager.java
	core/java/android/content/pm/IPackageManager.aidl
	core/java/android/net/NetworkScorerAppManager.java
	core/java/android/provider/Settings.java
	core/java/android/view/KeyEvent.java
	core/java/android/webkit/WebViewFactory.java
	core/jni/com_android_internal_os_Zygote.cpp
	core/res/AndroidManifest.xml
	core/res/res/layout/resolver_list_with_default.xml
	core/res/res/values/attrs.xml
	core/res/res/values/config.xml
	core/res/res/values/symbols.xml
	packages/DocumentsUI/src/com/android/documentsui/DirectoryFragment.java
	packages/DocumentsUI/src/com/android/documentsui/LocalPreferences.java
	packages/Keyguard/src/com/android/keyguard/EmergencyButton.java
	packages/SystemUI/res/layout-sw600dp/navigation_bar.xml
	packages/SystemUI/res/layout/system_icons.xml
	packages/SystemUI/res/values/config.xml
	packages/SystemUI/src/com/android/systemui/BatteryMeterView.java
	packages/SystemUI/src/com/android/systemui/ImageWallpaper.java
	packages/SystemUI/src/com/android/systemui/qs/QSPanel.java
	packages/SystemUI/src/com/android/systemui/qs/QSTileView.java
	packages/SystemUI/src/com/android/systemui/qs/tiles/CellularTile.java
	packages/SystemUI/src/com/android/systemui/qs/tiles/HotspotTile.java
	packages/SystemUI/src/com/android/systemui/qs/tiles/LocationTile.java
	packages/SystemUI/src/com/android/systemui/qs/tiles/RotationLockTile.java
	packages/SystemUI/src/com/android/systemui/recents/Recents.java
	packages/SystemUI/src/com/android/systemui/recents/RecentsActivity.java
	packages/SystemUI/src/com/android/systemui/recents/views/RecentsView.java
	packages/SystemUI/src/com/android/systemui/recents/views/TaskStackView.java
	packages/SystemUI/src/com/android/systemui/statusbar/BaseStatusBar.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/KeyguardStatusBarView.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/NavigationBarView.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/NotificationPanelView.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBarPolicy.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/QSTileHost.java
	packages/SystemUI/src/com/android/systemui/statusbar/policy/BatteryController.java
	packages/SystemUI/src/com/android/systemui/statusbar/policy/KeyButtonView.java
	packages/SystemUI/src/com/android/systemui/statusbar/tv/TvStatusBar.java
	services/backup/java/com/android/server/backup/BackupManagerService.java
	services/core/java/com/android/server/AppOpsService.java
	services/core/java/com/android/server/LockSettingsService.java
	services/core/java/com/android/server/SystemConfig.java
	services/core/java/com/android/server/am/ActivityManagerService.java
	services/core/java/com/android/server/am/ActivityStackSupervisor.java
	services/core/java/com/android/server/am/ProcessRecord.java
	services/core/java/com/android/server/audio/AudioService.java
	services/core/java/com/android/server/content/AppIdleMonitor.java
	services/core/java/com/android/server/content/SyncManager.java
	services/core/java/com/android/server/content/SyncStorageEngine.java
	services/core/java/com/android/server/display/AutomaticBrightnessController.java
	services/core/java/com/android/server/input/InputManagerService.java
	services/core/java/com/android/server/job/controllers/AppIdleController.java
	services/core/java/com/android/server/net/NetworkPolicyManagerService.java
	services/core/java/com/android/server/notification/NotificationManagerService.java
	services/core/java/com/android/server/pm/DefaultPermissionGrantPolicy.java
	services/core/java/com/android/server/pm/PackageManagerService.java
	services/core/java/com/android/server/pm/UserManagerService.java
	services/core/java/com/android/server/policy/PhoneWindowManager.java
	services/core/java/com/android/server/power/ShutdownThread.java
	services/core/java/com/android/server/wm/WindowManagerService.java
	services/core/jni/com_android_server_input_InputManagerService.cpp
	services/java/com/android/server/SystemServer.java
	services/net/java/android/net/dhcp/DhcpClient.java
	services/usage/java/com/android/server/usage/UsageStatsService.java
	services/usage/java/com/android/server/usage/UserUsageStatsService.java
	telephony/java/android/telephony/SubscriptionInfo.java
	telephony/java/com/android/ims/ImsCallProfile.java

+++ services/core/java/com/android/server/pm/PackageInstallerService.java
@@ -140,6 +140,7 @@ public class PackageInstallerService extends IPackageInstaller.Stub {
     private static final String ATTR_APP_ICON = "appIcon";
     private static final String ATTR_APP_LABEL = "appLabel";
     private static final String ATTR_ORIGINATING_URI = "originatingUri";
+    private static final String ATTR_ORIGINATING_UID = "originatingUid";
     private static final String ATTR_REFERRER_URI = "referrerUri";
     private static final String ATTR_ABI_OVERRIDE = "abiOverride";
     private static final String ATTR_VOLUME_UUID = "volumeUuid";
@@ -156,7 +157,6 @@ public class PackageInstallerService extends IPackageInstaller.Stub {
     private final PackageManagerService mPm;
 
     private AppOpsManager mAppOps;
-    private StorageManager mStorage;
 
     private final HandlerThread mInstallThread;
     private final Handler mInstallHandler;
@@ -212,14 +212,15 @@ public class PackageInstallerService extends IPackageInstaller.Stub {
         mCallbacks = new Callbacks(mInstallThread.getLooper());
 
         mSessionsFile = new AtomicFile(
-                new File(Environment.getSystemSecureDirectory(), "install_sessions.xml"));
-        mSessionsDir = new File(Environment.getSystemSecureDirectory(), "install_sessions");
+                new File(Environment.getDataSystemDirectory(), "install_sessions.xml"));
+        mSessionsDir = new File(Environment.getDataSystemDirectory(), "install_sessions");
         mSessionsDir.mkdirs();
 
         synchronized (mSessions) {
             readSessionsLocked();
 
-            reconcileStagesLocked(StorageManager.UUID_PRIVATE_INTERNAL);
+            reconcileStagesLocked(StorageManager.UUID_PRIVATE_INTERNAL, false /*isEphemeral*/);
+            reconcileStagesLocked(StorageManager.UUID_PRIVATE_INTERNAL, true /*isEphemeral*/);
 
             final ArraySet<File> unclaimedIcons = newArraySet(
                     mSessionsDir.listFiles());
@@ -240,11 +241,10 @@ public class PackageInstallerService extends IPackageInstaller.Stub {
 
     public void systemReady() {
         mAppOps = mContext.getSystemService(AppOpsManager.class);
-        mStorage = mContext.getSystemService(StorageManager.class);
     }
 
-    private void reconcileStagesLocked(String volumeUuid) {
-        final File stagingDir = buildStagingDir(volumeUuid);
+    private void reconcileStagesLocked(String volumeUuid, boolean isEphemeral) {
+        final File stagingDir = buildStagingDir(volumeUuid, isEphemeral);
         final ArraySet<File> unclaimedStages = newArraySet(
                 stagingDir.listFiles(sStageFilter));
 
@@ -258,18 +258,14 @@ public class PackageInstallerService extends IPackageInstaller.Stub {
         for (File stage : unclaimedStages) {
             Slog.w(TAG, "Deleting orphan stage " + stage);
             synchronized (mPm.mInstallLock) {
-                if (stage.isDirectory()) {
-                    mPm.mInstaller.rmPackageDir(stage.getAbsolutePath());
-                } else {
-                    stage.delete();
-                }
+                mPm.removeCodePathLI(stage);
             }
         }
     }
 
     public void onPrivateVolumeMounted(String volumeUuid) {
         synchronized (mSessions) {
-            reconcileStagesLocked(volumeUuid);
+            reconcileStagesLocked(volumeUuid, false /*isEphemeral*/);
         }
     }
 
@@ -310,12 +306,12 @@ public class PackageInstallerService extends IPackageInstaller.Stub {
     }
 
     @Deprecated
-    public File allocateStageDirLegacy(String volumeUuid) throws IOException {
+    public File allocateStageDirLegacy(String volumeUuid, boolean isEphemeral) throws IOException {
         synchronized (mSessions) {
             try {
                 final int sessionId = allocateSessionIdLocked();
                 mLegacySessions.put(sessionId, true);
-                final File stageDir = buildStageDir(volumeUuid, sessionId);
+                final File stageDir = buildStageDir(volumeUuid, sessionId, isEphemeral);
                 prepareStageDir(stageDir);
                 return stageDir;
             } catch (IllegalStateException e) {
@@ -386,8 +382,8 @@ public class PackageInstallerService extends IPackageInstaller.Stub {
         final int sessionId = readIntAttribute(in, ATTR_SESSION_ID);
         final int userId = readIntAttribute(in, ATTR_USER_ID);
         final String installerPackageName = readStringAttribute(in, ATTR_INSTALLER_PACKAGE_NAME);
-        final int installerUid = readIntAttribute(in, ATTR_INSTALLER_UID,
-                mPm.getPackageUid(installerPackageName, userId));
+        final int installerUid = readIntAttribute(in, ATTR_INSTALLER_UID, mPm.getPackageUid(
+                installerPackageName, PackageManager.MATCH_UNINSTALLED_PACKAGES, userId));
         final long createdMillis = readLongAttribute(in, ATTR_CREATED_MILLIS);
         final String stageDirRaw = readStringAttribute(in, ATTR_SESSION_STAGE_DIR);
         final File stageDir = (stageDirRaw != null) ? new File(stageDirRaw) : null;
@@ -405,6 +401,8 @@ public class PackageInstallerService extends IPackageInstaller.Stub {
         params.appIcon = readBitmapAttribute(in, ATTR_APP_ICON);
         params.appLabel = readStringAttribute(in, ATTR_APP_LABEL);
         params.originatingUri = readUriAttribute(in, ATTR_ORIGINATING_URI);
+        params.originatingUid =
+                readIntAttribute(in, ATTR_ORIGINATING_UID, SessionParams.UID_UNKNOWN);
         params.referrerUri = readUriAttribute(in, ATTR_REFERRER_URI);
         params.abiOverride = readStringAttribute(in, ATTR_ABI_OVERRIDE);
         params.volumeUuid = readStringAttribute(in, ATTR_VOLUME_UUID);
@@ -477,6 +475,7 @@ public class PackageInstallerService extends IPackageInstaller.Stub {
         writeStringAttribute(out, ATTR_APP_PACKAGE_NAME, params.appPackageName);
         writeStringAttribute(out, ATTR_APP_LABEL, params.appLabel);
         writeUriAttribute(out, ATTR_ORIGINATING_URI, params.originatingUri);
+        writeIntAttribute(out, ATTR_ORIGINATING_UID, params.originatingUid);
         writeUriAttribute(out, ATTR_REFERRER_URI, params.referrerUri);
         writeStringAttribute(out, ATTR_ABI_OVERRIDE, params.abiOverride);
         writeStringAttribute(out, ATTR_VOLUME_UUID, params.volumeUuid);
@@ -674,7 +673,9 @@ public class PackageInstallerService extends IPackageInstaller.Stub {
             File stageDir = null;
             String stageCid = null;
             if ((params.installFlags & PackageManager.INSTALL_INTERNAL) != 0) {
-                stageDir = buildStageDir(params.volumeUuid, sessionId);
+                final boolean isEphemeral =
+                        (params.installFlags & PackageManager.INSTALL_EPHEMERAL) != 0;
+                stageDir = buildStageDir(params.volumeUuid, sessionId, isEphemeral);
             } else {
                 stageCid = buildExternalStageCid(sessionId);
             }
@@ -773,12 +774,15 @@ public class PackageInstallerService extends IPackageInstaller.Stub {
         throw new IllegalStateException("Failed to allocate session ID");
     }
 
-    private File buildStagingDir(String volumeUuid) {
+    private File buildStagingDir(String volumeUuid, boolean isEphemeral) {
+        if (isEphemeral) {
+            return Environment.getDataAppEphemeralDirectory(volumeUuid);
+        }
         return Environment.getDataAppDirectory(volumeUuid);
     }
 
-    private File buildStageDir(String volumeUuid, int sessionId) {
-        final File stagingDir = buildStagingDir(volumeUuid);
+    private File buildStageDir(String volumeUuid, int sessionId, boolean isEphemeral) {
+        final File stagingDir = buildStagingDir(volumeUuid, isEphemeral);
         return new File(stagingDir, "vmdl" + sessionId + ".tmp");
     }
 
@@ -858,19 +862,25 @@ public class PackageInstallerService extends IPackageInstaller.Stub {
                 IntentSender statusReceiver, int userId) {
         final int callingUid = Binder.getCallingUid();
         mPm.enforceCrossUserPermission(callingUid, userId, true, true, "uninstall");
+        boolean allowSilentUninstall = true;
         if ((callingUid != Process.SHELL_UID) && (callingUid != Process.ROOT_UID)) {
             mAppOps.checkPackage(callingUid, callerPackageName);
+            final String installerPackageName = mPm.getInstallerPackageName(packageName);
+            allowSilentUninstall = mPm.isOrphaned(packageName) ||
+                    (installerPackageName != null
+                            && installerPackageName.equals(callerPackageName));
         }
 
-        // Check whether the caller is device owner
+        // Check whether the caller is device owner, in which case we do it silently.
         DevicePolicyManager dpm = (DevicePolicyManager) mContext.getSystemService(
                 Context.DEVICE_POLICY_SERVICE);
-        boolean isDeviceOwner = (dpm != null) && dpm.isDeviceOwnerApp(callerPackageName);
+        boolean isDeviceOwner = (dpm != null) && dpm.isDeviceOwnerAppOnCallingUser(
+                callerPackageName);
 
         final PackageDeleteObserverAdapter adapter = new PackageDeleteObserverAdapter(mContext,
                 statusReceiver, packageName, isDeviceOwner, userId);
-        if (mContext.checkCallingOrSelfPermission(android.Manifest.permission.DELETE_PACKAGES)
-                == PackageManager.PERMISSION_GRANTED) {
+        if (allowSilentUninstall && mContext.checkCallingOrSelfPermission(
+                android.Manifest.permission.DELETE_PACKAGES) == PackageManager.PERMISSION_GRANTED) {
             // Sweet, call straight through!
             mPm.deletePackage(packageName, adapter.getBinder(), userId, flags);
         } else if (isDeviceOwner) {
@@ -896,7 +906,10 @@ public class PackageInstallerService extends IPackageInstaller.Stub {
         mContext.enforceCallingOrSelfPermission(android.Manifest.permission.INSTALL_PACKAGES, TAG);
 
         synchronized (mSessions) {
-            mSessions.get(sessionId).setPermissionsResult(accepted);
+            PackageInstallerSession session = mSessions.get(sessionId);
+            if (session != null) {
+                session.setPermissionsResult(accepted);
+            }
         }
     }
